refactor: parse source service
